NP-16765 Move handle DoiRequest buttons to action panel
